業務日誌のデータベース、新規相談記録のバリデーション、編集、削除のセキュリティー対策を修正、追加

+++ app/Http/Controllers/RecordsController.php
@@ -52,6 +52,7 @@ class RecordsController extends Controller
     {
         // 入力された値の検証
         $this->validate($request, [
+            'content' => 'required',
             'image' => [
                 'file',
                 'mimes:jpeg,jpg,png'
@@ -92,11 +93,11 @@ class RecordsController extends Controller
      */
     public function show(Request $request, $id, $record_id)
     {
-        // 加工した記録番号の取得
-        $index = $request->input('index');
         // 注目する利用者とその相談記録の情報を取得
         $patient = Patient::find($id);
         $record = Record::find($record_id);
+        // 加工した記録番号の取得
+        $index = $request->input('index');
         
         // view の呼び出し
         return view('records.show', compact('patient', 'record', 'index'));
@@ -110,12 +111,20 @@ class RecordsController extends Controller
      */
     public function edit(Request $request, $id, $record_id)
     {
-        $index = $request->input('index');
         // 注目する利用者とその相談記録の情報を取得
         $patient = Patient::find($id);
         $record = Record::find($record_id);
-        // view の呼び出し
-        return view('records.edit', compact('patient', 'record', 'index'));
+        // 加工した記録番号の取得
+        $index = $request->input('index');
+        
+        // 注目している利用者の相談記録がログインしている職員が登録したものならば
+        if($record->user->id === \Auth::id()){
+            // view の呼び出し
+            return view('records.edit', compact('patient', 'record', 'index'));
+        } else {
+            // ログイン前のトップ画面へリダイレクト
+            return redirect('index');
+        }
     }
 
     /**
@@ -130,46 +139,54 @@ class RecordsController extends Controller
         // 注目している利用者とその相談記録の情報を取得
         $patient = Patient::find($id);
         $record = Record::find($record_id);
-        
         // 加工した記録番号の取得
         $index = $request->input('index');
         
-        // 入力値の検証
-        $this->validate($request, [
-            'image' => [
-                'file',
-                'mimes:jpeg,jpg,png'
-            ]
-        ]);
-        
-        // 入力情報の取得
-        $patient_id = $request->input('patient_id');
-        $content = $request->input('content');
-        $file =  $request->image;
-        
-        // 画像のアップロード
-        // https://qiita.com/ryo-program/items/35bbe8fc3c5da1993366
-        if($file){
-            // 現在時刻ともともとのファイル名を組み合わせてランダムなファイル名作成
-            $image = time() . $file->getClientOriginalName();
-            // アップロードするフォルダ名取得
-            $target_path = public_path('uploads/');
-            // アップロード処理
-            $file->move($target_path, $image);
-        }else{
-            // 画像が選択されていなければ、元の画像名のまま
-            $image = $record->image;
+        // 注目している利用者の相談記録がログインしている職員が登録したものならば
+        if($record->user->id === \Auth::id()) {
+            // 入力値の検証
+            $this->validate($request, [
+                'content' => 'required',
+                'image' => [
+                    'file',
+                    'mimes:jpeg,jpg,png'
+                ]
+            ]);
+            
+            // 入力情報の取得
+            $patient_id = $request->input('patient_id');
+            $content = $request->input('content');
+            $file =  $request->image;
+            
+            // 画像のアップロード
+            // https://qiita.com/ryo-program/items/35bbe8fc3c5da1993366
+            if($file){
+                // 現在時刻ともともとのファイル名を組み合わせてランダムなファイル名作成
+                $image = time() . $file->getClientOriginalName();
+                // アップロードするフォルダ名取得
+                $target_path = public_path('uploads/');
+                // アップロード処理
+                $file->move($target_path, $image);
+            }else{
+                // 画像が選択されていなければ、元の画像名のまま
+                $image = $record->image;
+            }
+            
+            // 入力情報をもとにrecordインスタンス情報を更新
+            $record->content = $content;
+            $record->image = $image;
+            
+            // データベースを更新
+            $record->save();
+           
+            // 注目する利用者の相談記録一覧へリダイレクト
+            return redirect('/patients/' . $patient_id . '/records')->with('flash_message', $patient->name . 'の記録番号: ' . $index . 'を更新しました');
+            
+        } else {
+            // ログイン前のトップ画面へリダイレクト
+            return redirect('index');
         }
         
-        // 入力情報をもとにrecordインスタンス情報を更新
-        $record->content = $content;
-        $record->image = $image;
-        
-        // データベースを更新
-        $record->save();
-       
-        // トップページへリダイレクト
-        return redirect('/patients/' . $patient_id . '/records')->with('flash_message', $patient->name . 'の記録番号: ' . $index . 'を更新しました');
     }
 
     /**
@@ -184,14 +201,20 @@ class RecordsController extends Controller
         $patient = Patient::find($id);
         $record = Record::find($record_id);
         $patient_id = $request->input('patient_id');
-        
         // 加工した記録番号の取得
         $index = $request->input('index');
         
-        // その相談記録の情報をデータベースから削除
-        $record->delete($record_id);
-        // リダイレクト
-        return redirect('/patients/' . $patient_id . '/records')->with('flash_message', $patient->name . 'の相談記録を1件削除しました');
+        // 注目している利用者の相談記録がログインしている職員が登録したものならば
+        if($record->user->id === \Auth::id()) {
+            // その相談記録の情報をデータベースから削除
+            $record->delete($record_id);
+            // リダイレクト
+            return redirect('/patients/' . $patient_id . '/records')->with('flash_message', $patient->name . 'の相談記録を1件削除しました');
+        } else {
+            // ログイン前のトップ画面へリダイレクト
+            return redirect('index');
+        }
+        
     }
     
     // 相談記録のキーワード検索
